+++ pdp/static/js/prism_demo_map.js
@@ -53,21 +53,32 @@ function init_prism_map() {
         }
     );
 
-    $('#map-title').html(params.layers + '<br />' + ncwms.params.TIME);
     getNCWMSLayerCapabilities(ncwms); // async save into global var ncwmsCapabilities
     current_dataset = params.layers;
 
-    function customize_wms_params(layer_name) {
-	var varname = layer_name.split('/')[1];
-	if (varname == 'pr') {
-	    this.params.LOGSCALE = true;
-	    this.params.STYLES = 'boxfill/occam_inv';
-	} else {
-	    this.params.LOGSCALE = false;
-	    this.params.STYLES = 'boxfill/ferret';
-	}
+    var set_map_title = function (layer_name) {
+        // 'this' must be bound to the ncwms layer object
+        var d = new Date(this.params.TIME);
+        if( layer_name.match(/_yr_/) ) { // is yearly
+            var date = d.getFullYear();
+        } else {
+            var date = d.getFullYear() + '/' + (d.getMonth() + 1);
+        }
+        $('#map-title').html(layer_name + '<br />' + date);
+
+        return true;
     };
-    ncwms.events.register('change', ncwms, customize_wms_params);
+
+    function ncwms_params(layer_name) {
+        var varname = layer_name.split('/')[1];
+        if (varname == 'pr') {
+            this.params.LOGSCALE = true;
+            this.params.STYLES = 'boxfill/occam_inv';
+        } else {
+            this.params.LOGSCALE = false;
+            this.params.STYLES = 'boxfill/ferret';
+        }
+    };
 
     map.addLayers(
         [
@@ -86,6 +97,29 @@ function init_prism_map() {
     map.getSelectionLayer = function() {
         return map.getLayersByName(selLayerName)[0];
     }
+
+    var cb = new Colorbar("pdpColorbar", ncwms);
+    cb.refresh_values();
+
+    ncwms.events.registerPriority('change', ncwms, function (layer_id) {
+        var params = {
+            id: layer_id.split('/')[0],
+            var: layer_id.split('/')[1]
+        }
+        var metadata_req = $.ajax(
+        {
+            url: "../metadata.json?request=GetMinMaxWithUnits",
+            data: params
+        });
+        metadata_req.done(function(data) {
+            var new_params = ncwms_params.call(ncwms, layer_id);
+            ncwms.mergeNewParams(new_params); // this does a layer redraw
+            cb.force_update(data.min, data.max, data.units) // must be called AFTER ncwms params updated
+        });
+    });
+
+    ncwms.events.register('change', ncwms, set_map_title)
+    ncwms.events.triggerEvent('change', defaults.dataset + "/" + defaults.variable);
 
     return map
 };